Singluarities works in most cases. Add caching of functions.
Curves f3 and f5 result in all zero singularity data. Perhaps
this has to do with the RootOf construct.

Caching is added to several functions, including
* puiseux.py:puiseux()
* integralbasis.py:integral_basis()
* singularities.py:singularities()

Since several of these functions are used often, with the same
arguments, it is best to cache some fo these results. Puiseux
can still take a little bit of time when symbolic coefficients
are involved.