Fix minor bug in monodromy and checkpoint homology code.
In using Muller's algorithm for complex root finding
we encounter a divide by zero error when dy = 0. We fix
this by adding an mpmath.eps "error" to dy when
computing guesses.

We also checkpoint homology in deleting / ignoring
the constuction of a- and b- cycles since we can simplify
future computations by just integrating over the c-cycles
and then taking appropriate linear combinations.