Replace simplify() with powsimp() in puiseux.py:base_expand
For performance purposes, we perform a powsimp() instead of
a full simplify. This is because we know that the final
expression should be a Puiseux series in T. Note that this
method does not simplify the coefficients of T all that
well. However, it does result in on average better
performance when running integral_basis().