Improve performance of Puiseux.
The calculation of Puiseux series is drastically improved by
using the combinatorial equations to compute the
coefficients of the series as opposed to using various
combinations of sympy.expand() and sympy.simplify(). Some
additional work needs to be done to decide how to normalize
these coefficients when, say, calling intbasis.valuation().