Add close root detection to monodromy.
Monodromy numerically approximates roots of discriminant
polynomials for speed purposes. Also, exact results are not
necessary since the paths stay far enough away from the
discriminant points. However, a root with multiplicity
is sometimes approximated as two distinct roots.

We use sympy.mpmath.almosteq() to detect close roots. Note
that if the two roots are _indeed_ supposed to be
distinct then the exepected genus of the curve may
be lower than the actual genus.