Refactor monodromy code.
It only made partial sense to have Monodromy be a class
that is inherited by RiemannSurface since, by the time
monodromy is computed, we have all of the information
we need to pass on to RiemannSurface. So, for simplicity,
we refactor the code so it behaves like a function,
instead. This should make it easier to use
RiemannSurfacePath in the future for performing the
analytic continuation.