Monodromy refactor.
We refactor monodromy to behave like a function.
Additionally, we eliminate the need to compute an "initial
monodromy" and simply use the combinatorial data to
determine whether or not a given monodromy path should
pass above or below a branch point.

We also use RiemannSurfacePath objects to perform the
analytic continuation.