Fix branch point comparison tool.
In the case when two or more branch points are colinear with the base
point of the monodromy group the branch points closest to the base
point occur "before" those further away. This fix produces monodromy
group matching those of Maple. (When using the same base point and
base sheets ordering.)