Fix infinity path bug.
The positive direction around infinity is a circle around the fininte
branch points in teh negative direction. A sign error abou this
property was resolved.

Now, f2 produces a period matrix. It's still not confirmed whether
or not this period matrix is a period matrix of f2 but it at least
satisfies the properties of a Riemann matrix.

f10 is being tested but, unfortunately, does not produce a valid period
matrix.