Naive integration of Cythonic Smale to paths.
Now that all of the Smale's alpha theory is Cythonized we
naively integrate it into RiemannSurfacePath. That is, we
perform no Cythonization of data types withing RSPath other
than the necessary one: RSPath.df, the y-derivatives of
the curve.

The next step will be to incrementally improve performance
within the class, itself.