Implement Abel Map and Discriminant Paths.
RiemannSurfacePaths and AnalyticContinuators to discriminant
places are implemented. Integration of paths is handled by
AnalyticContinuators since the strategy decision for
integration is the same as the strategy decision for
evaluation.

An AbelMap_Function class is created with a preliminary
implementation of the map. Additional (minor) work needs
to be done to allow for arbitrary starting places.