Implement new technique for computing Puiseux.
Use the method described in Poteaux and Rybowicz [1] for computing the
singular part of a Puiseux series. Use a modular Newton iteration
technique for computing the regular parts of the series. These new
techniques are implemented so as to properly handle series with RootOf
coefficients as well as for performance improvements.

Note that the PuiseuxTSeries objects have not yet been written to use
the new output data.

[1] "Towards a Symbolic-Numeric Method to Compute Puiseux Series : The
Modular Part", Adrien Poteaux and Marc Rybowicz,
http://arxiv.org/abs/0803.3027v1