Separate Differential numerator and denominator
We separately store the numerator and denominator of a
Differential for both display and computational purposes.
Every Abelian differential of the first kind has dfdy as
a numerator, yet, in the cases when the numerator
cancels with part of the denominator this is not exactly
clear.

Also, for computing the valuation divisor of such a
differential we want to get at the numerator separately
since the discriminant points have already been computed
in RiemannSurface.