Add caching to AbelMap
It is expected that telatively few places will be used
when evaluating the AbelMap. Therefore, it would be good
to cache the primitive evaluations in _eval_primitive().

We choose not to cache based on top-level input because,
even though there may be more divisors sent to AbelMap,
these divisors often consist of the same places. It
doesn't make sense to "double cache" these values if the
only difference is in a simple linear algebra computation.