Puiseux returns correct branching / multiplicity
pusieux.pusieux() did not return the correct branching
number or multiplicitiy. This is because the Poteaux /
Rybowicz algorithm only returns a single Puiseux series
for a single root of the Netwon polygon polynomial, not
for each root.

This was a fix made a long time ago but, of course, was not
kept in various refactors.