Initial second kind differentials
Begin the sturcture behind constructing Abelian differentials of
the second kind.

The main application of the second kind differentials is the
calculation of the coefficient vectors in KP solutions. However,
it turns out that simply constructing such differentials is
difficult. (Involves the prime form, etc.) An alternate technique
involving the first kind differentials is used instead.

For now, we pause on developing the second kind differentials