Raise meaningful RootOf error message in pusieux
Due to current (version 0.7.7) limitations in Sympy, roots of
polynomials with non-rational coefficients cannot be constructed
thus limiting the places where Puiseux series can be computed.
Specifically, RootOfs cannot be computed when the polynommial
in question contains RootOf coefficients.

Instead of raising a meaningless NotImplementedError about roots
of EX a NotImplementedError describing this situation is raised.