Improve cache decorators
The new function and instancemethod (class method) caching
decorator now supports signature and documentation forwarding.
This makes the reference section of the documentation properly
identify the functions.

The downside is that these caching decorators are slower. This
should not be a problem with the kinds of computations they're
applied to.