Update TravisCI configuration to use Sage
To ensure the best compatibility with different platforms we install
Abelfunctions into Sage using the old-style .spkg method. From
experiments it seems that compilation of Abelfunctions outside of
Sage (that is, sans .spkg) only works when the user has a source
distribution of Sage.

In the future a separate repo should contain the spkg version so
as to not confuse users.