Replace Sphinx documentation with simple Markdown
We remove the old Sphinx documentation and replace with a collection
of Markdown files. Currently, there is no easy way to automatically
compile the Sphinx documentation on ReadTheDocs since Sage is
required to compile the code.

The most helpful documentation at the moment is probably the tutorial
collections. In the future we will need some way of auto-generating and
presenting the software reference.