Update symmetric homology notebook
This notebook is an experimental implementation of the symmetric
homology basis algorithm by C. Kalla and C. Klein. The current version
works for all given H and Q examples.

However, the genus 6 period matrix might be resulting in an incorrect N1
matrix because the algorithm fails to find a suitable row below the
diagonal with which to eliminate. There exists a suitable row above the
diagonal but this is "off-limits" in the sense that it will disrupt the
block diagonal structure already computed.

Perhaps this is still an error in the QH algorithm.