Compute Z-basis and N1 matrix
The N1 matrix is a matrix over GF(2) from which we determine the
topological type of the Riemann surface. It requires computing a Z-basis
of the integer kernel of R^T - I_{2g} where R is the anti-holomorphic
involution matrix.

Two routines for computing these items are given as well as a basic
necessary, but not sufficient, test for the validity of N1.