Fix transposition bug
The entire Kalla and Klein algorithm actually works the transposes of
the a- and b-periods. (This is not mentioned anywhere in the paper.)
This fact confirms why the examples in the paper do not satisfy the
property that A^{-1}B is a Riemann matrix.

The entry function `symmetrize_periods` transforms the input and output
to the appropriate form.

Add final pieces of documentation.