Use correct method to obtain Numpy include dirs
The setup.py script needs to grab the directories containing the Numpy
include files in order to compile many of the Cython modules. We switch
to using Numpy's `get_numpy_include_dirs()` function for this purpose as
opposed to a manual technique which caused the build system to break
when the Numpy directory changed.