[#134] Update finite_sum to C99 standard
Once nice feature of the C99 standard is the ability to create
dynamically-sized stack-allocated arrays. These perform much better than
heap-allocated arrays and, since we only need to allocate temporary
vectors of length equal to the genus we needn't worry about exceeding
stack size limits.

We use this feature in the finite_sum.c functions as well as clean up
some of the formatting and documentation.