[#137] Remove module-level matplotlib imports
The module-level imports of the matplotlib library causes long delays /
hangs in the TravisCI tests due to an issue regarding the construction
of a font cache. Since the plotting of Abelfunctions is not tested these
imports are moved into the functions and methods that require matplotlib
for displaying images.

A longer term solution to this problem would be to move all plotting
routines to use Sage's built-in plotting. See #138.