Delete riemann_surface_path.py
This file is not actually used because `riemann_surface_path.pyx` exists. I'm guessing it was kept either accidentaly or to compare against the `.pyx` version but I think it's very confusing when trying to search/browse the codebase and of course a burden to maintain if the goal is to keep them in sync.